:sparkles: [feat] 잔액 데이터 타입 롤백 #S12P31E205-108

+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/fintech/domain/Account.java
@@ -23,7 +23,7 @@ public class Account {
     private String bankCode;
 
     @Column(nullable = false)
-    private Long balance;
+    private Integer balance;
 
     @Column(name = "created_at", nullable = false)
     private LocalDateTime createdAt;
