+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/fintech/domain/RegistAccount.java
@@ -44,7 +44,7 @@ public class RegistAccount {
     private Boolean primary;
 
     @Column(nullable = false)
-    private Long balance;
+    private Integer balance;
 
     @Column(name = "registered_at", nullable = false)
     private LocalDateTime registeredAt;
